error output now compatible with vim quickfix
Prior to this change, read errors and parse errors would cause vim quickfix to
open new buffers with meaningless paths. For logged output, vim would open a new
buffer with the timestamp as a name.

This change makes some assumptions that are true today:
 * Errors from ioutil.ReadFile are user-readable and meaningful alone.
    e.g. open example.go: no such file or directory
 * Errors from parsing concatenate with the filename to produce quickfix format.
    e.g. example.go:8:1: expected declaration, found '}'

With this change, the suggested autosave behavior from the README is usable.