lint: fix if/else if/else if/else return handling
If you have multiple "else if" statements and the last one has
a return, golint will incorrectly warn that the last else can be
dedented, even if the if/first else if blocks don't.

From the issue report: "the problem is that the ignore flag is not
getting propagated to all the else's in a chain because of the early
return." Flipping the order fixes the problem.

Fixes golang/lint#354. Patch was suggested by Ramesh Vyaghrapuri, who
also identified the problem.