lint: allow multiple error return values
When returning multiple values with an error,
lint checks if the error is the last return value.
But the implementation actually is checking for all return values
except for the last one, and throw the alert if it found an error.

There is a (edge) case where some function returning more than one error
is getting a false positive even when the last return value is an error.

This patch adds an early check, to see if the last return value is an error
and if so, it will pass silently.

Fixes golang/lint#286

Change-Id: Ib3ad50ed127cdab6ac63e9d89c1c97f4a641f972
GitHub-Last-Rev: 9dd9de4fce5fccc846d010cc1c57720e33da85aa
GitHub-Pull-Request: golang/lint#387
Reviewed-on: https://go-review.googlesource.com/99295
Reviewed-by: Andrew Bonventre <andybons@golang.org>