cmd/golint: Add support for structured output of lints
It's helpful when running golint in any type of automated setting
to have a structured output of the lints. That way, when storing
your golint output, it's easy to handle the data without needing
to go through the process of parsing the output which is never fun.

This adds a flag to the program `-json`. When used, it will provide
all of the data encapsulated in the Problem type.

Here is an example invocation run from the lint directory of this
project:

```
$ golint -json -min_confidence .1 | jq .
[
  {
    "Position": {
      "Filename": "import.go",
      "Offset": 0,
      "Line": 1,
      "Column": 1
    },
    "Text": "should have a package comment, unless it's in another file for this package",
    "Link": "https://golang.org/wiki/CodeReviewComments#package-comments",
    "Confidence": 0.2,
    "LineText": "package main\n",
    "Category": "comments",
    "ReplacementLine": ""
  }
]
```