go/lint: only report kBlah error for constants and top level vars
The error for variables that start with k followed by a variable name,
such as "kBlah", has a lot of false positives. It's intended to avoid
cases where users try to follow the Google C++ style guide or hungarian
notation for constants. So it should only be reported for constants and
top-level variables because those are the cases where users make this
mistake.

Change-Id: I7e862dbc1013707b6b9ebaa72ee07fb623dd0ce8
Reviewed-on: https://go-review.googlesource.com/c/154339
Reviewed-by: Heschi Kreinick <heschi@google.com>
Reviewed-by: Rebecca Stambler <rstambler@golang.org>
Run-TryBot: Heschi Kreinick <heschi@google.com>
TryBot-Result: Gobot Gobot <gobot@golang.org>