lint: do not suggest keywords as Go names
In some circumstances, programmers are tempted to use Go reserved
keywords as names (e.g. "type" or "default"). As this is forbidden
in Go, the compiler will reject it. A common workaround is to append
an underline to the keyword to make the compiler happy, e.g.,
changing the name "type" to "type_".

Previously golint considered that use of underline bad style and
wrongly suggested employing the original keyword instead:

astVarRef.go:6:2: don't use underscores in Go names; struct field type_ should be type
cmdIf.go:32:3: don't use underscores in Go names; var goto_ should be goto

This CL changes golint so that it doesn't make that invalid suggestion.

Fixes golang/lint#393